test(consultant-playground): split physical_boundary.rs by concern
physical_boundary.rs was at 1192/1200 lines against the file-size gate.
Split into physical_boundary_manifest.rs (Cargo metadata boundary) and
physical_boundary_source.rs (source-shape boundary), with shared lexer,
JSON parser, and fixture helpers factored into tests/support/*.rs and
pulled in per binary via #[path] so each test crate only compiles what
it actually calls. Same 8 tests, same assertions, no behavior change.
